Move actions into own subdirectory
