+++ python/lsst/analysis/tools/actions/keyedData/__init__.py
@@ -1,0 +1,2 @@
+from .keyedDataActions import *
+from .stellarLocusFit import *
