+++ python/lsst/analysis/tools/actions/scalar/__init__.py
