--- python/lsst/analysis/tools/keyedDataActions/__init__.py
@@ -1,2 +1,0 @@
-from .keyedDataActions import *
-from .stellarLocusFit import StellarLocusFitAction
